Initial Solution State, showing only a class library. Q: What do we test? Q: How do we test?

+++ 00_SolutionUnderTest/SUT.FinancialCalculator/IBonusCalculator.cs
@@ -4,6 +4,6 @@ namespace SUT.FinancialCalculator
 {
     public interface IBonusCalculator
     {
-        decimal GetBonusPercentage(Employee employee);
+        double GetBonusPercentage(Employee employee);
     }
 }
